+++ app/src/main/kotlin/com/maltaisn/notes/receiver/AlarmReceiver.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Nicolas Maltais
+ * Copyright 2022 Nicolas Maltais
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,6 +20,7 @@ import android.app.PendingIntent
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
+import android.os.Build
 import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationManagerCompat
 import com.maltaisn.notes.App
@@ -89,14 +90,18 @@ class AlarmReceiver : BroadcastReceiver() {
 
         // Add actions for non-recurring reminders
         if (note.reminder?.recurrence == null) {
+            var pendingIntentFlags = PendingIntent.FLAG_UPDATE_CURRENT
+            if (Build.VERSION.SDK_INT >= 23) {
+                pendingIntentFlags = pendingIntentFlags or PendingIntent.FLAG_IMMUTABLE
+            }
+
             // Mark done action
             val markDoneIntent = Intent(context, AlarmReceiver::class.java).apply {
                 action = ACTION_MARK_DONE
                 putExtra(EXTRA_NOTE_ID, noteId)
             }
             builder.addAction(R.drawable.ic_check, context.getString(R.string.action_mark_as_done),
-                PendingIntent.getBroadcast(context, noteId.toInt(), markDoneIntent,
-                    PendingIntent.FLAG_UPDATE_CURRENT))
+                PendingIntent.getBroadcast(context, noteId.toInt(), markDoneIntent, pendingIntentFlags))
 
             // Postpone action only if not recurring.
             val postponeIntent = Intent(context, NotificationActivity::class.java).apply {
@@ -106,8 +111,7 @@ class AlarmReceiver : BroadcastReceiver() {
             }
             builder.addAction(R.drawable.ic_calendar,
                 context.getString(R.string.action_postpone),
-                PendingIntent.getActivity(context, noteId.toInt(), postponeIntent,
-                     PendingIntent.FLAG_UPDATE_CURRENT))
+                PendingIntent.getActivity(context, noteId.toInt(), postponeIntent, pendingIntentFlags))
         }
 
         NotificationManagerCompat.from(context).notify(noteId.toInt(), builder.build())